Fix bug in code
Found bug in that the command line input was being converted to integer and not a bigger data type like long or float and subsequently fixed that.

+++ 0x0C-more_malloc_free/101-mul.c
@@ -97,8 +97,8 @@ int main(int argc, char **argv)
 	}
 	else if (b && c)
 	{
-		d = atoi(argv[1]);
-		e = atoi(argv[2]);
+		d = atol(argv[1]);
+		e = atol(argv[2]);
 	}
 	product = d * e;
 	printf("%lu\n", product);
